chore: move error handling to core.errors
- Move core.config.errors to core.errors
- Move utils.handle_exception to core.errors
