+++ ggshield/cmd/secret/scan/ci.py
@@ -9,8 +9,9 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
     create_output_handler,
 )
 from ggshield.core.cache import ReadOnlyCache
+from ggshield.core.errors import handle_exception
 from ggshield.core.git_shell import check_git_dir, get_list_commit_SHA
-from ggshield.core.utils import EMPTY_SHA, handle_exception
+from ggshield.core.utils import EMPTY_SHA
 from ggshield.scan import ScanContext, ScanMode
 from ggshield.scan.repo import scan_commit_range
