+++ ggshield/cmd/secret/scan/path.py
@@ -8,9 +8,9 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
     create_output_handler,
 )
 from ggshield.core.constants import MAX_WORKERS
+from ggshield.core.errors import handle_exception
 from ggshield.core.file_utils import get_files_from_paths
 from ggshield.core.text_utils import create_progress_bar
-from ggshield.core.utils import handle_exception
 from ggshield.scan import ScanCollection, ScanContext, ScanMode, SecretScanner
