+++ ggshield/cmd/secret/scan/prereceive.py
@@ -12,9 +12,10 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
     create_output_handler,
 )
 from ggshield.core.cache import ReadOnlyCache
+from ggshield.core.errors import handle_exception
 from ggshield.core.git_shell import get_list_commit_SHA, git
 from ggshield.core.text_utils import display_error
-from ggshield.core.utils import EMPTY_SHA, PRERECEIVE_TIMEOUT, handle_exception
+from ggshield.core.utils import EMPTY_SHA, PRERECEIVE_TIMEOUT
 from ggshield.output import GitLabWebUIOutputHandler
 from ggshield.output.text.message import remediation_message
 from ggshield.scan import ScanContext, ScanMode
