+++ ggshield/cmd/secret/scan/range.py
@@ -6,8 +6,8 @@ from ggshield.cmd.secret.scan.secret_scan_common_options import (
     add_secret_scan_common_options,
     create_output_handler,
 )
+from ggshield.core.errors import handle_exception
 from ggshield.core.git_shell import get_list_commit_SHA
-from ggshield.core.utils import handle_exception
 from ggshield.scan import ScanContext, ScanMode
 from ggshield.scan.repo import scan_commit_range
