+++ ggshield/core/config/user_config.py
@@ -7,7 +7,6 @@ import click
 import marshmallow_dataclass
 from marshmallow import ValidationError
 
-from ggshield.core.config.errors import ParseError, format_validation_error
 from ggshield.core.config.utils import (
     get_global_path,
     load_yaml_dict,
@@ -20,6 +19,7 @@ from ggshield.core.constants import (
     GLOBAL_CONFIG_FILENAMES,
     LOCAL_CONFIG_PATHS,
 )
+from ggshield.core.errors import ParseError, format_validation_error
 from ggshield.core.types import FilteredConfig, IgnoredMatch, IgnoredMatchSchema
 from ggshield.core.utils import api_to_dashboard_url
 from ggshield.iac.policy_id import POLICY_ID_PATTERN, validate_policy_id
