+++ ggshield/core/utils.py
@@ -1,7 +1,6 @@
 import logging
 import os
 import re
-import traceback
 from datetime import datetime
 from enum import Enum
 from typing import Iterable, List, NamedTuple, Optional
@@ -11,7 +10,6 @@ import click
 from dotenv import load_dotenv
 from pygitguardian.models import Match
 
-from ggshield.core.config.errors import ExitCode, _ExitError
 from ggshield.core.constants import ON_PREMISE_API_URL_PATH_PREFIX
 
 from .git_shell import get_git_root, is_git_dir
@@ -218,22 +216,6 @@ json_output_option_decorator = click.option(
     show_default=True,
     help="JSON output results",
 )
-
-
-def handle_exception(e: Exception, verbose: bool) -> int:
-    """
-    Handle exception from a scan command.
-    """
-    if isinstance(e, click.exceptions.Abort):
-        return ExitCode.SUCCESS
-    exit_code = e.exit_code if isinstance(e, _ExitError) else ExitCode.UNEXPECTED_ERROR
-
-    if not isinstance(e, click.ClickException) and verbose:
-        traceback.print_exc()
-
-    display_error(str(e))
-
-    return exit_code
 
 
 def _find_dot_env() -> Optional[str]:
