+++ tests/unit/cmd/test_config.py
@@ -5,7 +5,7 @@ import pytest
 
 from ggshield.cmd.main import cli
 from ggshield.core.config import Config
-from ggshield.core.config.errors import ExitCode
+from ggshield.core.errors import ExitCode
 
 from .utils import prepare_config
