+++ tests/unit/core/test_utils.py
@@ -9,7 +9,7 @@ from pygitguardian import GGClient
 from ggshield.core.cache import Cache
 from ggshield.core.client import create_client_from_config
 from ggshield.core.config import Config
-from ggshield.core.config.errors import UnknownInstanceError
+from ggshield.core.errors import UnknownInstanceError
 from ggshield.core.utils import (
     MatchIndices,
     api_to_dashboard_url,
